Cleaned up code to pass throught parameters and reamove DRY code
