+++ client/src/pages/Types.ts
@@ -8,3 +8,16 @@ export type Progress = {
   level: number;
   progress: number;
 };
+
+export type LayoutProps = {
+  moods: Mood[];
+  progress?: Progress;
+  selectEmoji?: number;
+  counter: string;
+  isOpen?: boolean;
+  handleSelectedEmoji: (index: number) => void;
+  handleCharacterCount: (event: React.ChangeEvent<HTMLTextAreaElement>) => void;
+  handleSubmit: (event: React.FormEvent<HTMLFormElement>) => void;
+  openModal?: () => void;
+  closeModal?: () => void;
+};
